+++ Roguelike/src/org/tilegames/hexicube/topdownproto/entity/EntityRat.java
@@ -33,15 +33,15 @@ public class EntityRat extends EntityLiving
 	@Override
 	public double damageAfterResistance(double damage, DamageType type)
 	{
-		if(type == DamageType.ACID) return (long) (damage * 1.5);
+		if(type == DamageType.ACID) return damage * 1.5;
 		if(type == DamageType.BLUNT) return damage * 2;
 		if(type == DamageType.CRUSHING) return damage * 3;
 		if(type == DamageType.EXPLOSIVE) return damage * 7;
 		if(type == DamageType.FALLING) return damage;
-		if(type == DamageType.FIRE) return (long) (damage * 0.4);
+		if(type == DamageType.FIRE) return damage * 0.4;
 		if(type == DamageType.GENERIC) return damage;
-		if(type == DamageType.ICE) return (long) (damage * 0.4);
-		if(type == DamageType.SHARP) return (long) (damage * 0.75);
+		if(type == DamageType.ICE) return damage * 0.4;
+		if(type == DamageType.SHARP) return damage * 0.75;
 		return 0;
 	}
